refactor(vault): move event types into module
